tbl/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
codacy-coverage==1.3.6
coverage==4.4.1
Django==1.11.3
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
idna==2.5
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
py==1.4.34
PyJWT==1.5.2
pytest==3.1.3
pytest-cov==2.5.1
pytest-django==3.1.2
pytz==2017.2
requests==2.18.2
urllib3==1.22
django-taggit-0.22.1
